Refactor aggregators
