+++ server/aggregators/aggregator-page.js
@@ -1,15 +1,17 @@
+var db = require('../database');
 
-module.exports = function(db, query, callback) {
-    db.collection('logs').aggregate([
-        {$match: {'hash.scriptHash': {$eq: query.id}}},
+module.exports = function(conn, query, callback) {
+    db.aggregate(conn, [
+        {$match: {'hash.pageHash': {$eq: query.id}}},
         {$group: {
             _id: '$hash.messageHash',
             id: {$first: '$hash.messageHash'},
             title: {$first: '$message'},
             count: {$sum: 1},
+            browsers: {$addToSet: '$ua.family'},
             earliest: {$min: '$timestamp'},
             latest: {$max: '$timestamp'},
-            browsers: {$addToSet: '$ua.family'}
-        }}
+        }},
+        {$sort: {count: -1}}
     ], callback);
 };
