+++ server/aggregators/aggregator-script.js
@@ -1,16 +1,16 @@
+var db = require('../database');
 
-module.exports = function(db, query, callback) {
-    db.collection('logs').aggregate([
-        {$match: {'hash.pageHash': {$eq: query.id}}},
+module.exports = function(conn, query, callback) {
+    db.aggregate(conn, [
+        {$match: {'hash.scriptHash': {$eq: query.id}}},
         {$group: {
             _id: '$hash.messageHash',
             id: {$first: '$hash.messageHash'},
             title: {$first: '$message'},
             count: {$sum: 1},
-            browsers: {$addToSet: '$ua.family'},
             earliest: {$min: '$timestamp'},
             latest: {$max: '$timestamp'},
-        }},
-        {$sort: {count: -1}}
+            browsers: {$addToSet: '$ua.family'}
+        }}
     ], callback);
 };
