+++ server/aggregators/aggregator-scripts.js
@@ -1,11 +1,13 @@
+var db = require('../database');
 
-module.exports = function(db, query, callback) {
+module.exports = function(conn, query, callback) {
     var scriptTitle = {$cond: {
         if: {$ifNull: ['$line', false]},
         then: {$concat: ['$url', ':', '$line']},
         else: {$concat: ['$url', ':', '0']}
     }};
-    db.collection('logs').aggregate([
+
+    db.aggregate(conn, [
         {$group: {
             _id: '$hash.scriptHash',
             id: {$first: '$hash.scriptHash'},
