+++ server/aggregators/compose-sha.js
