+++ server/aggregators/message-signature.js
