+++ server/route-reports.js
@@ -1,4 +1,4 @@
-var aggregators = require('./dbutils/aggregators');
+var aggregators = require('./aggregators');
 
 module.exports = function (req, res, next) {
     var type = req.params.type;
@@ -7,11 +7,9 @@ module.exports = function (req, res, next) {
     if (aggregator) {
         aggregator(req._db, req.query, function(err, result) {
             if (err) {
-                console.log(err);
                 res.status(500).end();
                 return next();
             }
-            console.log(result);
             res.json(result);
             next();
         });
